+++ src/docs/devguide/hugo/site/content/release-notes.md
@@ -1,5 +1,10 @@
 This page describes changes to the Service Interface Base API implemented across different releases of the interface.
 
+## 3.24.0
+Released 30 April 2020
+* Added Account PaymentMethod.
+* Added Interfaces for HasAmounts & HasPaymentMethods Polymorphic Utility Reuse
+
 ## 3.23.0
 Released 25 November 2019
 * Added utility methods to JsonUtil class to assist in reading the contents of a file as a string and deserialising JSON objects from files.